# src/main/resources/openapi.yml
openapi: '3.0.2'
info:
  title: Project Management API
  description: API for managing projects
  version: '1.0'
servers:
  - url: http://localhost:8080
...
paths:
...
  /collaborator/{projectId}:
    get:
      tags:
        - Collaborator
      summary: Get all the collaborators of a specific project.
      operationId: getCollaboratorsByProjectId
      description: Returns the collaborators based on the provided Project ID.
      parameters:
        - name: projectId
          in: path
          description: ID of the Project whose collaborators to retrieve.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collaborator'

    post:
      tags:
        - Collaborator
      summary: Add a new collaborator to a Project.
      operationId: addCollaboratorToProject
      description: Adds the collaborator based on Project Id and returns it.
      parameters:
        - name: projectId
          in: path
          description: ID of the Project whose collaborator to add.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Collaborator'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collaborator'
  /collaborator/{collaboratorId}:
    put:
      tags:
        - Collaborator
      summary: Edit a collaborator of the Plaform.
      operationId: editCollaboratorOfProject
      description: Edits the collaborator based on their Id and returns it.
      parameters:
        - name: collaboratorId
          in: path
          description: ID of the collaborator to edit.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Collaborator'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collaborator'

    delete:
      tags:
        - Collaborator
      summary: Delete a collaborator of the Plaform.
      operationId: deleteCollaborator
      description: Delete a collaborator based on their Id.
      parameters:
        - name: collaboratorId
          in: path
          description: ID of the collaborator to edit.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collaborator'
  /collaborator/{projectId}/{collaboratorId}:
    delete:
      tags:
        - Collaborator
      summary: Delete a collaborator.
      operationId: deleteCollaboratorFromProject
      description: Deletes a collaborator based on their and project's Id.
      parameters:
        - name: collaboratorId
          in: path
          description: ID of the collaborator to be deleted.
          required: true
          schema:
            type: string
            format: uuid
        - name: projectId
          in: path
          description: ID of the project where the deletion happens.
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collaborator'
#Request
  /request:
    get:
      tags:
        - Request
      summary: Get all requests in the system
      operationId: getRequests
      description: Retrieve all requests in the database.
      responses:
        '200':
          description: Requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
        '401':
          description: Unauthorized - user is not authorized to retrieve requests
  /request/{username}:
    get:
      tags:
        - Request
      summary: Get requests for a user
      operationId: getRequestsForUser
      description: Retrieve all requests associated with a user
      parameters:
        - name: username
          in: path
          description: username of the user for which the requests have to be found.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
        '401':
          description: Unauthorized - user is not authorized to retrieve requests
        '404':
          description: User not found
  /request/{projectId}:

    get:
      tags:
        - Request
      summary: Get requests for a project
      operationId: getRequestsForProject
      description: Retrieve all requests associated with a project
      parameters:
        - name: projectId
          in: path
          description: ID of the project for which requests are being retrieved
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Requests retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Request'
        '401':
          description: Unauthorized - user is not authorized to retrieve requests
        '404':
          description: Project not found

    put:
      tags:
        - Request
      summary: Create a request
      operationId: createRequest
      description: Create a new request for a project
      parameters:
        - name: projectId
          in: path
          description: ID of the project for which the request is being created
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request'
      responses:
        '201':
          description: Request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '400':
          description: Bad request - invalid input data
        '401':
          description: Unauthorized - user is not authorized to create a request
        '404':
          description: Project not found
  /request/{requestId}:
    delete:
      tags:
        - Request
      summary: Delete a request
      operationId: deleteRequest
      description: Delete a request
      parameters:
        - name: requestId
          in: path
          description: ID of the request to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Request deleted successfully
        '401':
          description: Unauthorized - user is not authorized to delete the request
        '404':
          description: Request not found
components:
  schemas:
...
    Request:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        newTitle:
          type: string
        newDescription:
          type: string
        newBibtex:
          type: string
        isCounterOffer:
          type: boolean
...
    Collaborator:
      type: object
      properties:
        collaboratorId:
          type: string
          format: uuid
        name:
          type: string